[clean] Seperating fast, local, unit tests from longer, application-driven, functional tests
