+++ tests/unit/firmware/fatstacks/test_screen_FullScreen.py
