+++ tests/unit/firmware/test_structs_Firmware.py
