+++ tests/unit/firmware/test_versions_VersionManager.py
