+++ tests/unit/test_error_ApplicationError.py
